==> lecture21-random-forest/00_download_model_fits.ipynb <==
### This notebook is to download and unzip the fitted models used in  `02_random_forest_airbnb.ipynb`.

Since the cross-validation and model fits take a lots of time, we saved the results and uploaded them to an OSF repository which you can find [here](https://osf.io/mw4xj/?view_only=). You can also download the `model_fits.zip` folder from the repository and extract the zip file by hand, but the following code chunk will also do this for you.

import requests, zipfile
from io import BytesIO

print("Downloading started")
url = "https://osf.io/nsa3q/download"
filename = "model_fits.zip"

# Downloading the file by sending the request to the URL
req = requests.get(url)
print("Downloading completed")

# extracting the zip file contents
zipfile = zipfile.ZipFile(BytesIO(req.content))
zipfile.extractall(".")
print("Folder unzipped")
